\name{HardinRockeExtension-package}
\alias{HardinRockeExtension-package}
\alias{HardinRockeExtension}
\docType{package}
\title{
  Package for replicating the results of Hardin and Rocke (2005)
  and Green and Martin (2014).
}
\description{
  Package for replicating the results of Hardin and Rocke (2005)
  and Green and Martin (2014).
}
\details{
\tabular{ll}{
Package: \tab HardinRockeExtension\cr
Type: \tab Package\cr
Version: \tab 1.0.8\cr
Date: \tab 2015-12-08\cr
License: \tab GPL-2\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Written and maintained by Christopher G. Green <christopher.g.green@gmail.com>
}
\references{
J. Hardin and D. M. Rocke. The distribution of robust distances. 
Journal of Computational and Graphical Statistics, 14:928-946, 2005.
}
\keyword{ package }
\keyword{ robust }
\keyword{ multivariate }
\seealso{
  \code{\link[CerioliOutlierDetection:CerioliOutlierDetection-package]{CerioliOutlierDetection}} 
}
\examples{
  \dontrun{
      hrSim(p=5,n=50  ,N=5000,B=500)
  }
}
